Add protect with roles add validations to tasks endpoints

+++ apps/eclass-website/pages/api/v1/task/[id].ts
@@ -1,8 +1,14 @@
+import { Role } from "@prisma/client";
 import type { NextApiResponse } from "next";
 import { protect } from "../../../../middleware/protect";
 import { reqWithUser } from "../../../../types/reqWithUser";
+import getUserCourses from "../../../../utils/getUserCourses";
+import getUserOwnedCourses from "../../../../utils/getUserOwnedCourses";
+import userIsInCourse from "../../../../utils/userIsInCourse";
+import userOwnsCourse from "../../../../utils/userOwnsCourse";
 
 const handler = async (req: reqWithUser, res: NextApiResponse) => {
+  const unauthorized = res.status(401).json({success: false, message: "Unauthorized"});
   switch (req.method) {
     case "GET":
       return getTaskById();
@@ -23,8 +29,19 @@ const handler = async (req: reqWithUser, res: NextApiResponse) => {
         id: req.query.id.toString(),
       },
     });
-
+    
+    
     if (task) {
+      if (
+        !(await userIsInCourse(req.user.id, task.courseId) ||
+          await userOwnsCourse(req.user.id, task.courseId))
+        ){
+        return res.status(401).json({
+          success: false,
+          message: "Usuario no pertenece al curso de la tarea ni es dueño del curso"
+        })
+      }
+      
       return res.status(200).json({
         success: true,
         task: task,
@@ -38,18 +55,34 @@ const handler = async (req: reqWithUser, res: NextApiResponse) => {
   }
   /// updates an task given an task in the body of the request
   async function updateTask() {
+    if (req.user.role === Role.student) return unauthorized;
     if (req.body) {
       try {
-        const task = await prisma.task.update({
+        const task = await prisma.task.findUnique({
+          where: {
+            id: req.query.id.toString(),
+          },
+        });
+
+        if (task){
+          if (!await userOwnsCourse(req.user.id, task.courseId)){
+            return res.status(401).json({
+              success: false,
+              message: "El usuario no es dueño de este curso"
+            })
+          }
+        }
+
+        const updatedTask = await prisma.task.update({
           where: {
             id: req.query.id.toString(),
           },
           data: req.body,
         });
-        if (task) {
+        if (updatedTask) {
           return res.status(200).json({
             success: true,
-            task: task,
+            task: updatedTask,
           });
         }
       } catch (error) {
@@ -63,15 +96,31 @@ const handler = async (req: reqWithUser, res: NextApiResponse) => {
 
   // deletes an task given an id
   async function deleteTask() {
+    if (req.user.role === Role.student) return unauthorized;
     try {
-      const task = await prisma.task.delete({
+      const task = await prisma.task.findUnique({
+        where: {
+          id: req.query.id.toString(),
+        },
+      });
+
+      if (task){
+        if (!await userOwnsCourse(req.user.id, task.courseId)){
+          return res.status(401).json({
+            success: false,
+            message: "El usuario no es dueño de este curso"
+          })
+        }
+      }
+
+      const deletedTask = await prisma.task.delete({
         where: {
           id: req.query.id.toString(),
         },
       });
       return res.status(200).json({
         success: true,
-        task: task,
+        task: deletedTask,
       });
     } catch (error) {
       return res.status(400).json({
